refactor(toolbar): fold popover content into sheet views
